Make response assertions less strict
To make the response assertions less brittle, I'd suggest that we only assert that the expected keys exist and not the exact set of expected keys was returned. This allows the device to return more fields if it wants to so that the library continues working if keys are added in firmware updates. It still allows us to be aware of breaking changes to the API.

:D