Do not bind / listen on same port that we send UDP datagrams to
It turns out that devices are pretty much able to handle random sender
port for discovery.

If a library would be extended to act as a slave in grouped we might
need to change this code to listen to specific port (7777).

Primary motivation for this change is to be able to run multiple
instances of CLI/library at the same time.