UDP client: do not try to compare received datagrams to own addresses
It turned out that own_addresses() didn't return anything if hostname
was defined in /etc/hosts within 127.0.0.0/8.

We no longer listen to a port that we are sending to, we won't
receive our own datagrams.

If the library would be extended in a way that we could act as slave
grouped device and we might end up listening to same port as we are
sending, this code might need to be changed to not detect our packets
again.