Withold tokens for known unauthenticated REST API endpoints in control
This should slightly speed up calls to firmware_version() and
device_info() if we didn't have authentication token to pass as we don't
need to get it.

As a side consequence this should break infinite recursion when we need
to verify login with devices' hardware address which we might not have
at the moment the call and we can't call device_info() to get device's
mac.