Explicitly require known click version for use by black
Recent `black` linter runs stopped working and produced following
traceback:

    Traceback (most recent call last):
      File "[...]/xled/.tox/black/bin/black", line 8, in <module>
        sys.exit(patched_main())
      File "[...]/xled/.tox/black/lib/python3.10/site-packages/black/__init__.py", line 1372, in patched_main
        patch_click()
      File "[...]/home/pbabinca/local/workspace/xled/.tox/black/lib/python3.10/site-packages/black/__init__.py", line 1358, in patch_click
        from click import _unicodefun
    ImportError: cannot import name '_unicodefun' from 'click' ([...]/xled/.tox/black/lib/python3.10/site-packages/click/__init__.py)

According to https://stackoverflow.com/a/71674345/83400 :

    This has been fixed by Black 22.3.0. Versions before that won't work with
    click 8.1.0.

But we already pinned to last `black` version (21.12b0) that still supports
python2 syntax and therefore we cannot upgrade. This change pins known `click`
version that works with `black` and it has to stay like this until we support
python2.